requirements.txt
django==4.2.0
djangorestframework==3.14.0
django-filter==23.2
django-cors-headers==4.1.0
celery==5.3.0
redis==4.5.5
psycopg2-binary==2.9.6
beautifulsoup4==4.12.2
requests==2.31.0
feedparser==6.0.10
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.5.0
django-celery-beat==2.5.0
django-celery-results==2.5.0